+++ src/bot.js
@@ -1,16 +1,9 @@
 const Eris = require('eris');
 const config = require('./config');
 
-const bot = new Eris.CommandClient(config.token, {
+const bot = new Eris.Client(config.token, {
   getAllUsers: true,
-}, {
-  prefix: config.prefix,
-  ignoreSelf: true,
-  ignoreBots: true,
-  defaultHelpCommand: false,
-  defaultCommandOptions: {
-    caseInsensitive: true,
-  },
+  restMode: true,
 });
 
 module.exports = bot;