+++ src/data/threads.js
@@ -1,15 +1,21 @@
-const Eris = require('eris');
+const {User, Member} = require('eris');
+
 const transliterate = require('transliteration');
 const moment = require('moment');
 const uuid = require('uuid');
 const humanizeDuration = require('humanize-duration');
 
+const bot = require('../bot');
 const knex = require('../knex');
 const config = require('../config');
 const utils = require('../utils');
+const updates = require('./updates');
 
 const Thread = require('./Thread');
 const {THREAD_STATUS} = require('./constants');
+
+const MINUTES = 60 * 1000;
+const HOURS = 60 * MINUTES;
 
 /**
  * @param {String} id
@@ -45,24 +51,63 @@ function getHeaderGuildInfo(member) {
 
 /**
  * Creates a new modmail thread for the specified user
- * @param {Eris.User} user
+ * @param {User} user
+ * @param {Member} member
  * @param {Boolean} quiet If true, doesn't ping mentionRole or reply with responseMessage
- * @returns {Promise<Thread>}
+ * @returns {Promise<Thread|undefined>}
  * @throws {Error}
  */
-async function createNewThreadForUser(user, quiet = false) {
+async function createNewThreadForUser(user, quiet = false, ignoreRequirements = false) {
   const existingThread = await findOpenThreadByUserId(user.id);
   if (existingThread) {
     throw new Error('Attempted to create a new thread for a user with an existing open thread!');
   }
 
-  // Check the config for a requirement of account age to contact modmail,
-  // if the account is too young, return an optional message without making a new thread
-  if (config.requiredAccountAge) {
-    if (user.createdAt > moment() - config.requiredAccountAge * 3600000){
+  // If set in config, check that the user's account is old enough (time since they registered on Discord)
+  // If the account is too new, don't start a new thread and optionally reply to them with a message
+  if (config.requiredAccountAge && ! ignoreRequirements) {
+    if (user.createdAt > moment() - config.requiredAccountAge * HOURS){
       if (config.accountAgeDeniedMessage) {
         const privateChannel = await user.getDMChannel();
         await privateChannel.createMessage(config.accountAgeDeniedMessage);
+      }
+      return;
+    }
+  }
+
+  // Find which main guilds this user is part of
+  const mainGuilds = utils.getMainGuilds();
+  const userGuildData = new Map();
+
+  for (const guild of mainGuilds) {
+    let member = guild.members.get(user.id);
+
+    if (! member) {
+      try {
+        member = await bot.getRESTGuildMember(guild.id, user.id);
+      } catch (e) {
+        continue;
+      }
+    }
+
+    if (member) {
+      userGuildData.set(guild.id, { guild, member });
+    }
+  }
+
+  // If set in config, check that the user has been a member of one of the main guilds long enough
+  // If they haven't, don't start a new thread and optionally reply to them with a message
+  if (config.requiredTimeOnServer && ! ignoreRequirements) {
+    // Check if the user joined any of the main servers a long enough time ago
+    // If we don't see this user on any of the main guilds (the size check below), assume we're just missing some data and give the user the benefit of the doubt
+    const isAllowed = userGuildData.size === 0 || Array.from(userGuildData.values()).some(({guild, member}) => {
+      return member.joinedAt < moment() - config.requiredTimeOnServer * MINUTES;
+    });
+
+    if (! isAllowed) {
+      if (config.timeOnServerDeniedMessage) {
+        const privateChannel = await user.getDMChannel();
+        await privateChannel.createMessage(config.timeOnServerDeniedMessage);
       }
       return;
     }
@@ -78,10 +123,28 @@ async function createNewThreadForUser(user, quiet = false) {
 
   console.log(`[NOTE] Creating new thread channel ${channelName}`);
 
+  // Figure out which category we should place the thread channel in
+  let newThreadCategoryId;
+
+  if (config.categoryAutomation.newThreadFromGuild) {
+    // Categories for specific source guilds (in case of multiple main guilds)
+    for (const [guildId, categoryId] of Object.entries(config.categoryAutomation.newThreadFromGuild)) {
+      if (userGuildData.has(guildId)) {
+        newThreadCategoryId = categoryId;
+        break;
+      }
+    }
+  }
+
+  if (! newThreadCategoryId && config.categoryAutomation.newThread) {
+    // Blanket category id for all new threads (also functions as a fallback for the above)
+    newThreadCategoryId = config.categoryAutomation.newThread;
+  }
+
   // Attempt to create the inbox channel for this thread
   let createdChannel;
   try {
-    createdChannel = await utils.getInboxGuild().createChannel(channelName, null, 'Новый почтовый тред', config.newThreadCategoryId);
+    createdChannel = await utils.getInboxGuild().createChannel(channelName, null, 'Новый почтовый тред', newThreadCategoryId);
   } catch (err) {
     console.error(`Error creating modmail channel for ${user.username}#${user.discriminator}!`);
     throw err;
@@ -134,34 +197,40 @@ async function createNewThreadForUser(user, quiet = false) {
 
   let infoHeader = infoHeaderItems.join(', ');
 
-  // Guild info
-  const guildInfoHeaderItems = new Map();
-  const mainGuilds = utils.getMainGuilds();
+  // Guild member info
+  for (const [guildId, guildData] of userGuildData.entries()) {
+    const {nickname, joinDate} = getHeaderGuildInfo(guildData.member);
+    const headerItems = [
+      `НИКНЕЙМ **${utils.escapeMarkdown(nickname)}**`,
+      `ПРИСОЕДИНИЛСЯ **${joinDate}** назад`
+    ];
 
-  mainGuilds.forEach(guild => {
-    const member = guild.members.get(user.id);
-    if (! member) return;
-
-    const {nickname, joinDate} = getHeaderGuildInfo(member);
-    guildInfoHeaderItems.set(guild.name, [
-      `НИКНЕЙМ: **${nickname}**`,
-      `ПРИСОЕДИНИЛСЯ: **${joinDate}** назад`,
-      `\nРОЛИ: **${member.roles.length ? `@${member.roles.map(r => guild.roles.get(r)).sort((a, b) => b.position - a.position).map(r => r.name).join(', @')}` : 'нет'}**`
-    ]);
-  });
-
-  guildInfoHeaderItems.forEach((items, guildName) => {
-    if (mainGuilds.length === 1) {
-      infoHeader += `\n${items.join(', ')}`;
-    } else {
-      infoHeader += `\n**[${guildName}]** ${items.join(', ')}`;
+    if (guildData.member.voiceState.channelID) {
+      const voiceChannel = guildData.guild.channels.get(guildData.member.voiceState.channelID);
+      if (voiceChannel) {
+        headerItems.push(`ГОЛОСОВОЙ КАНАЛ **${utils.escapeMarkdown(voiceChannel.name)}**`);
+      }
     }
-  });
+
+    if (config.rolesInThreadHeader && guildData.member.roles.length) {
+      const roles = guildData.member.roles.map(roleId => guildData.guild.roles.get(roleId)).filter(Boolean);
+      headerItems.push(`РОЛИ **${roles.map(r => r.name).join(', ')}**`);
+    }
+
+    const headerStr = headerItems.join(', ');
+
+    if (mainGuilds.length === 1) {
+      infoHeader += `\n${headerStr}`;
+    } else {
+      infoHeader += `\n**[${utils.escapeMarkdown(guildData.guild.name)}]** ${headerStr}`;
+    }
+  }
 
   function declOfNum(n, titles) {
     return titles[(n % 10 == 1 && n % 100 != 11 ? 0 : n % 10 >= 2 && n % 10 <= 4 && (n % 100 < 10 || n % 100 >= 20) ? 1 : 2)];
   }
 
+  // ModMail history / previous logs
   const userLogCount = await getClosedThreadCountByUserId(user.id);
   if (userLogCount > 0) {
     infoHeader += `\n\nЭтот пользователь имеет **${userLogCount}** обращени${declOfNum(userLogCount, ['е', 'я', 'й'])}. Используйте \`${config.prefix}logs\` чтобы посмотреть их.`;
@@ -170,6 +239,13 @@ async function createNewThreadForUser(user, quiet = false) {
   infoHeader += '\n────────────────';
 
   await newThread.postSystemMessage(infoHeader);
+
+  if (config.updateNotifications) {
+    const availableUpdate = await updates.getAvailableUpdate();
+    if (availableUpdate) {
+      await newThread.postNonLogMessage(`📣 Доступна новая версия бота (${availableUpdate})`);
+    }
+  }
 
   // If there were errors sending a response to the user, note that
   if (responseMessageError) {