+++ src/modules/logs.js
@@ -1,52 +1,63 @@
-const threadUtils = require('../threadUtils');
 const threads = require("../data/threads");
 const moment = require('moment');
 const utils = require("../utils");
 
-module.exports = bot => {
-  const addInboxServerCommand = (...args) => threadUtils.addInboxServerCommand(bot, ...args);
+const LOG_LINES_PER_PAGE = 10;
 
-  addInboxServerCommand('logs', (msg, args, thread) => {
-    async function getLogs(userId) {
-      const userThreads = await threads.getClosedThreadsByUserId(userId);
+module.exports = (bot, knex, config, commands) => {
+  const logsCmd = async (msg, args, thread) => {
+    let userId = args.userId || (thread && thread.user_id);
+    if (! userId) return;
 
-      // Descending by date
-      userThreads.sort((a, b) => {
-        if (a.created_at > b.created_at) return -1;
-        if (a.created_at < b.created_at) return 1;
-        return 0;
-      });
+    let userThreads = await threads.getClosedThreadsByUserId(userId);
 
-      const threadLines = await Promise.all(userThreads.map(async thread => {
-        const logUrl = await thread.getLogUrl();
-        const formattedDate = moment.utc(thread.created_at).format('Do MMM [в] HH:mm [UTC]');
-        return `\`${formattedDate}\`: <${logUrl}>`;
-      }));
+    // Descending by date
+    userThreads.sort((a, b) => {
+      if (a.created_at > b.created_at) return -1;
+      if (a.created_at < b.created_at) return 1;
+      return 0;
+    });
 
-      const message = `**Логи <@${userId}>:**\n${threadLines.join('\n')}`;
+    // Pagination
+    const totalUserThreads = userThreads.length;
+    const maxPage = Math.ceil(totalUserThreads / LOG_LINES_PER_PAGE);
+    const inputPage = args.page;
+    const page = Math.max(Math.min(inputPage ? parseInt(inputPage, 10) : 1, maxPage), 1); // Clamp page to 1-<max page>
+    const isPaginated = totalUserThreads > LOG_LINES_PER_PAGE;
+    const start = (page - 1) * LOG_LINES_PER_PAGE;
+    const end = page * LOG_LINES_PER_PAGE;
+    userThreads = userThreads.slice((page - 1) * LOG_LINES_PER_PAGE, page * LOG_LINES_PER_PAGE);
 
-      // Send the list of logs in chunks of 15 lines per message
-      const lines = message.split('\n');
-      const chunks = utils.chunk(lines, 15);
+    const threadLines = await Promise.all(userThreads.map(async thread => {
+      const logUrl = await thread.getLogUrl();
+      const formattedDate = moment.utc(thread.created_at).format('Do MMM [в] HH:mm [UTC]');
+      return `\`${formattedDate}\`: <${logUrl}>`;
+    }));
 
-      let root = Promise.resolve();
-      chunks.forEach(lines => {
-        root = root.then(() => msg.channel.createMessage(lines.join('\n')));
-      });
+    let message = isPaginated
+      ? `**Логи <@${userId}>** (страница **${page}/${maxPage}**, лог **${start + 1}-${end}/${totalUserThreads}**):`
+      : `**Логи <@${userId}>:**`;
+
+    message += `\n${threadLines.join('\n')}`;
+
+    if (isPaginated) {
+      message += `\nЧтобы просмотреть дальше, добавьте номер страницы в конец команды`;
     }
 
-    if (args.length > 0) {
-      // User mention/id as argument
-      const userId = utils.getUserMention(args.join(' '));
-      if (! userId) return;
-      getLogs(userId);
-    } else if (thread) {
-      // Calling !logs without args in a modmail thread returns the logs of the user of that thread
-      getLogs(thread.user_id);
-    }
-  });
+    // Send the list of logs in chunks of 15 lines per message
+    const lines = message.split('\n');
+    const chunks = utils.chunk(lines, 15);
 
-  addInboxServerCommand('loglink', async (msg, args, thread) => {
+    let root = Promise.resolve();
+    chunks.forEach(lines => {
+      root = root.then(() => msg.channel.createMessage(lines.join('\n')));
+    });
+  };
+
+  commands.addInboxServerCommand('logs', '<userId:userId> [page:number]', logsCmd);
+  commands.addInboxServerCommand('logs', '[page:number]', logsCmd);
+
+  commands.addInboxServerCommand('loglink', [], async (msg, args, thread) => {
     if (! thread) {
       thread = await threads.findSuspendedThreadByChannelId(msg.channel.id);
       if (! thread) return;