+++ src/modules/snippets.js
@@ -2,10 +2,27 @@ const threads = require('../data/threads');
 const snippets = require('../data/snippets');
 const config = require('../config');
 const utils = require('../utils');
-const threadUtils = require('../threadUtils');
+const { parseArguments } = require('knub-command-manager');
 
-module.exports = bot => {
-  const addInboxServerCommand = (...args) => threadUtils.addInboxServerCommand(bot, ...args);
+const whitespaceRegex = /\s/;
+const quoteChars = ["'", '"'];
+
+module.exports = (bot, knex, config, commands) => {
+  /**
+   * "Renders" a snippet by replacing all argument placeholders e.g. {1} {2} with their corresponding arguments.
+   * The number in the placeholder is the argument's order in the argument list, i.e. {1} is the first argument (= index 0)
+   * @param {String} body
+   * @param {String[]} args
+   * @returns {String}
+   */
+  function renderSnippet(body, args) {
+    return body
+      .replace(/(?<!\\){\d+}/g, match => {
+        const index = parseInt(match.slice(1, -1), 10) - 1;
+        return (args[index] != null ? args[index] : match);
+      })
+      .replace(/\\{/g, '{');
+  }
 
   /**
    * When a staff member uses a snippet (snippet prefix + trigger word), find the snippet and post it as a reply in the thread
@@ -43,82 +60,75 @@ module.exports = bot => {
     const thread = await threads.findByChannelId(msg.channel.id);
     if (! thread) return;
 
-    const trigger = msg.content.replace(snippetPrefix, '').toLowerCase();
+    let [, trigger, rawArgs] = msg.content.slice(snippetPrefix.length).match(/(\S+)(?:\s+(.*))?/s);
+    trigger = trigger.toLowerCase();
+
     const snippet = await snippets.get(trigger);
     if (! snippet) return;
 
-    const replied = await thread.replyToUser(msg.member, snippet.body, [], isAnonymous);
+    let args = rawArgs ? parseArguments(rawArgs) : [];
+    args = args.map(arg => arg.value);
+    const rendered = renderSnippet(snippet.body, args);
+
+    const replied = await thread.replyToUser(msg.member, rendered, [], isAnonymous);
     if (replied) msg.delete();
   });
 
   // Show or add a snippet
-  addInboxServerCommand('snippet', async (msg, args, thread) => {
-    const trigger = args[0];
-    if (! trigger) return
-
-    const text = args.slice(1).join(' ').trim();
-    const snippet = await snippets.get(trigger);
+  commands.addInboxServerCommand('snippet', '<trigger> [text$]', async (msg, args, thread) => {
+    const snippet = await snippets.get(args.trigger);
 
     if (snippet) {
-      if (text) {
+      if (args.text) {
         // If the snippet exists and we're trying to create a new one, inform the user the snippet already exists
-        utils.postSystemMessageWithFallback(msg.channel, thread, `Шаблон "${trigger}" уже существует! Вы можете отредактировать или удалить его с помощью \`${config.prefix}edit_snippet\` и \`${config.prefix}delete_snippet\` соответственно.`);
+        utils.postSystemMessageWithFallback(msg.channel, thread, `Шаблон "${args.trigger}" уже существует! Вы можете отредактировать или удалить его с помощью \`${config.prefix}edit_snippet\` и \`${config.prefix}delete_snippet\` соответственно.`);
       } else {
         // If the snippet exists and we're NOT trying to create a new one, show info about the existing snippet
-        utils.postSystemMessageWithFallback(msg.channel, thread, `\`${config.snippetPrefix}${trigger}\` ответит:\n${snippet.body}`);
+        utils.postSystemMessageWithFallback(msg.channel, thread, `\`${config.snippetPrefix}${args.trigger}\` ответит: \`\`\`${utils.disableCodeBlocks(snippet.body)}\`\`\``);
       }
     } else {
-      if (text) {
+      if (args.text) {
         // If the snippet doesn't exist and the user wants to create it, create it
-        await snippets.add(trigger, text, msg.author.id);
-        utils.postSystemMessageWithFallback(msg.channel, thread, `Шаблон "${trigger}" создан!`);
+        await snippets.add(args.trigger, args.text, msg.author.id);
+        utils.postSystemMessageWithFallback(msg.channel, thread, `Шаблон "${args.trigger}" создан!`);
       } else {
         // If the snippet doesn't exist and the user isn't trying to create it, inform them how to create it
-        utils.postSystemMessageWithFallback(msg.channel, thread, `Шаблон "${trigger}" не существует! Вы можете создать его с помощью \`${config.prefix}snippet ${trigger} text\`.`);
-       }
+        utils.postSystemMessageWithFallback(msg.channel, thread, `Шаблон "${args.trigger}" не существует! Вы можете создать его с помощью \`${config.prefix}snippet ${args.trigger} текст\``);
+      }
     }
+  }, {
+    aliases: ['s']
   });
 
-  bot.registerCommandAlias('s', 'snippet');
-
-  addInboxServerCommand('delete_snippet', async (msg, args, thread) => {
-    const trigger = args[0];
-    if (! trigger) return;
-
-    const snippet = await snippets.get(trigger);
+  commands.addInboxServerCommand('delete_snippet', '<trigger>', async (msg, args, thread) => {
+    const snippet = await snippets.get(args.trigger);
     if (! snippet) {
-      utils.postSystemMessageWithFallback(msg.channel, thread, `Шаблон "${trigger}" не существует!`);
+      utils.postSystemMessageWithFallback(msg.channel, thread, `Шаблон "${args.trigger}" не существует!`);
       return;
     }
 
-    await snippets.del(trigger);
-    utils.postSystemMessageWithFallback(msg.channel, thread, `Шаблон "${trigger}" удален!`);
+    await snippets.del(args.trigger);
+    utils.postSystemMessageWithFallback(msg.channel, thread, `Шаблон "${args.trigger}" удален!`);
+  }, {
+    aliases: ['ds']
   });
 
-  bot.registerCommandAlias('ds', 'delete_snippet');
-
-  addInboxServerCommand('edit_snippet', async (msg, args, thread) => {
-    const trigger = args[0];
-    if (! trigger) return;
-
-    const text = args.slice(1).join(' ').trim();
-    if (! text) return;
-
-    const snippet = await snippets.get(trigger);
+  commands.addInboxServerCommand('edit_snippet', '<trigger> [text$]', async (msg, args, thread) => {
+    const snippet = await snippets.get(args.trigger);
     if (! snippet) {
-      utils.postSystemMessageWithFallback(msg.channel, thread, `Шаблон "${trigger}" не существует!`);
+      utils.postSystemMessageWithFallback(msg.channel, thread, `Шаблон "${args.trigger}" не существует!`);
       return;
     }
 
-    await snippets.del(trigger);
-    await snippets.add(trigger, text, msg.author.id);
+    await snippets.del(args.trigger);
+    await snippets.add(args.trigger, args.text, msg.author.id);
 
-    utils.postSystemMessageWithFallback(msg.channel, thread, `Шаблон "${trigger}" изменен!`);
+    utils.postSystemMessageWithFallback(msg.channel, thread, `Шаблон "${args.trigger}" изменен!`);
+  }, {
+    aliases: ['es']
   });
 
-  bot.registerCommandAlias('es', 'edit_snippet');
-
-  addInboxServerCommand('snippets', async (msg, args, thread) => {
+  commands.addInboxServerCommand('snippets', [], async (msg, args, thread) => {
     const allSnippets = await snippets.all();
     const triggers = allSnippets.map(s => s.trigger);
     triggers.sort();