+++ src/utils.js
@@ -65,10 +65,10 @@ function postLog(...args) {
   getLogChannel().createMessage(...args);
 }
 
-function postError(str) {
-  getLogChannel().createMessage({
-    content: `${getInboxMention()}**Ошибка:** ${str.trim()}`,
-    disableEveryone: false
+function postError(channel, str, opts = {}) {
+  return channel.createMessage({
+    ...opts,
+    content: `⚠ ${str}`
   });
 }
 
@@ -135,9 +135,11 @@ async function formatAttachment(attachment, attachmentUrl) {
  * @returns {String|null}
  */
 function getUserMention(str) {
+  if (! str) return null;
+
   str = str.trim();
 
-  if (str.match(/^[0-9]+$/)) {
+  if (isSnowflake(str)) {
     // User ID
     return str;
   } else {
@@ -224,7 +226,7 @@ const delayStringRegex = /^([0-9]+)(?:([dhms])[a-z]*)?/i;
 /**
  * Turns a "delay string" such as "1h30m" to milliseconds
  * @param {String} str
- * @returns {Number}
+ * @returns {Number|null}
  */
 function convertDelayStringToMS(str) {
   let match;
@@ -250,10 +252,15 @@ function convertDelayStringToMS(str) {
 }
 
 function getInboxMention() {
-  if (config.mentionRole == null) return '';
-  else if (config.mentionRole === 'here') return '@here ';
-  else if (config.mentionRole === 'everyone') return '@everyone ';
-  else return `<@&${config.mentionRole}> `;
+  const mentionRoles = Array.isArray(config.mentionRole) ? config.mentionRole : [config.mentionRole];
+  const mentions = [];
+  for (const role of mentionRoles) {
+    if (role == null) continue;
+    else if (role === 'here') mentions.push('@here');
+    else if (role === 'everyone') mentions.push('@everyone');
+    else mentions.push(`<@&${role}>`);
+  }
+  return mentions.join(' ') + ' ';
 }
 
 function postSystemMessageWithFallback(channel, thread, text) {
@@ -289,10 +296,19 @@ function setDataModelProps(target, props) {
 
 const snowflakeRegex = /^[0-9]{17,}$/;
 function isSnowflake(str) {
-  return snowflakeRegex.test(str);
+  return str && snowflakeRegex.test(str);
 }
 
 const humanizeDelay = (delay, opts = {}) => humanizeDuration(delay, Object.assign({conjunction: ' и ', language: 'ru'}, opts));
+
+const markdownCharsRegex = /([\\_*|`~])/g;
+function escapeMarkdown(str) {
+  return str.replace(markdownCharsRegex, '\\$1');
+}
+
+function disableCodeBlocks(str) {
+  return str.replace(/`/g, "`\u200b");
+}
 
 module.exports = {
   BotError,
@@ -327,4 +343,7 @@ module.exports = {
   isSnowflake,
 
   humanizeDelay,
+
+  escapeMarkdown,
+  disableCodeBlocks,
 };